add netlify plugin
