+++ src/pages/{StoreCollection.slug}/[...].tsx
@@ -1,27 +1,61 @@
 import { parseSearchParamsState } from '@vtex/store-sdk'
 import { graphql } from 'gatsby'
 import React, { useMemo } from 'react'
+import { priceRange } from 'src/sdk/search/priceRange'
 import View from 'src/views/collection'
+import type { SearchParamsState } from '@vtex/store-sdk'
 import type { PageProps } from 'gatsby'
 import type {
-  BrowserCollectionPageQueryQuery,
-  BrowserCollectionPageQueryQueryVariables,
-} from '@generated/BrowserCollectionPageQuery.graphql'
+  CollectionPageQueryQuery,
+  CollectionPageQueryQueryVariables,
+} from '@generated/CollectionPageQuery.graphql'
 
 export type Props = PageProps<
-  BrowserCollectionPageQueryQuery,
-  BrowserCollectionPageQueryQueryVariables & { slug: string }
+  CollectionPageQueryQuery,
+  CollectionPageQueryQueryVariables & { slug: string }
 >
 
-const useSearchParams = ({ href }: Location) =>
-  useMemo(() => href && parseSearchParamsState(new URL(href)), [href])
+const useSearchParams = (props: Props): SearchParamsState => {
+  return useMemo(() => {
+    const {
+      location: { href, pathname },
+      data: { storeCollection },
+      params: { '*': wildcard },
+    } = props
+
+    if (wildcard?.length > 0) {
+      return parseSearchParamsState(new URL(href))
+    }
+
+    const { selectedFacets } = storeCollection!.meta
+    const [base] = pathname.split(selectedFacets[0].value)
+
+    // TODO: Remove this bit of code once we have a graphql layer
+    const facets = selectedFacets.reduce((acc: any, facet: any) => {
+      const { key } = facet
+      const value =
+        key === 'priceRange'
+          ? priceRange.formatUrl(priceRange.parseQuery(facet.value)!)
+          : facet.value
+
+      acc.push({ key, value })
+
+      return acc
+    }, [] as SearchParamsState['selectedFacets'])
+
+    return {
+      page: 0,
+      base,
+      selectedFacets: facets,
+      term: null,
+      personalized: false,
+      sort: 'score-desc',
+    }
+  }, [props])
+}
 
 function Page(props: Props) {
-  const searchParams = useSearchParams(props.location)
-
-  if (!searchParams) {
-    return null
-  }
+  const searchParams = useSearchParams(props)
 
   return <View {...props} searchParams={searchParams} />
 }
@@ -30,7 +64,7 @@ function Page(props: Props) {
  * This query is run during SSG
  * */
 export const query = graphql`
-  query BrowserCollectionPageQuery($id: String!) {
+  query CollectionPageQuery($id: String!) {
     site {
       ...CollectionSeoFragment_site
     }
